Revert "refactor and db query update"
